requirements.txt
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
docopt==0.6.2
fonttools==4.37.1
idna==3.3
joblib==1.1.0
JPype1==1.4.0
kiwisolver==1.4.4
langid==1.1.6
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pipreqs==0.4.11
pyparsing==3.0.9
pyspellchecker==0.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.1
tweepy==4.10.1
urllib3==1.26.12
yarg==0.1.9
